requirement.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
umap-learn==0.5.5
sentence-transformers==2.6.1
joblib==1.4.2
python-dotenv==1.0.1
groq==0.5.0